cli tabled

+++ src/proto/from.rs
@@ -10,8 +10,8 @@ impl From<models::Namespace> for proto::Namespace {
             id: ns.id,
             name: ns.name,
             description: ns.description,
-            created: ns.created as u64,
-            modified: ns.modified as u64,
+            created: ns.created,
+            modified: ns.modified,
         }
     }
 }
@@ -22,8 +22,8 @@ impl From<proto::Namespace> for models::Namespace {
             id: ns.id,
             name: ns.name,
             description: ns.description,
-            created: ns.created as u128,
-            modified: ns.modified as u128,
+            created: ns.created,
+            modified: ns.modified,
         }
     }
 }
